[IMP] hr_attendance: say cheese!!! (and confirm your identity)
This commit adds
- "Take Pictures on Check-In" feature in Attendance. When enabled in company
settings, the employee's picture will be clicked when they check-in either
from "Systray" or using "Kiosk" (if camera is available in the Attendance device)
- a new confirmation page when capturing attendance through Kiosk's
Manual mode, in order to prevent attendances recorded by misclicks when
pin not enabled.

Key technical changes:
- extended attendance routes (systray, kiosk, manual selection) to handle
optional `check_in_image` parameter.
- stored captured images in new field `in_image` and added a log saying
"Check-in Image Capture"
- also, when we fail to capture the image; either because of large size
or camera unavailibility/permissions, a notification is sent to user.
- new components:
	- `AttendanceVideoStream` and a hook (`useCamera`) to start camera
	and display live video stream (preview) while trying to check-in using
	systray or in Kiosk mode.
	- `KioskConfirmation` in a flow of Manual attendance on Kiosk to act
	as a confirmation screen before the attendance is actually captured.
	(if pin not enabled)

task-5114645