[IMP] event: add multi entry for event tickets
This commit adds content mainly in event.registration to allow tickets that can be used multiple times by the buyers.
The registration is ongoing as long as there is remaining entries if there is multiple entries allowed on the ticket used by the registration.
Single use ticket are not affected by those changes and their behavior is the same as before.

Task-5373441