[FIX] mail, account: attachment delete access handling
Previously, if a user without write access to an attachment attempted to
delete it, a NotFound error was raised. Since this error is not handled
like other access-related errors, it appeared as a crash in the UI.

This commit replaces the NotFound error with an AccessError and disables
the delete button when the user does not have write access
to the attachment.

Task-4979444